requirements.txt
# Google API
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Slack
playwright

# Utilities
python-dotenv
requests
pyyaml
